IS: Move all queries for motestatusendringer to repository
