+++ src/test/kotlin/no/nav/syfo/dialogmote/api/v2/AvlysDialogmoteApiV2Spek.kt
@@ -21,7 +21,7 @@ import no.nav.syfo.brev.behandler.kafka.KafkaBehandlerDialogmeldingDTO
 import no.nav.syfo.brev.esyfovarsel.EsyfovarselProducer
 import no.nav.syfo.client.oppfolgingstilfelle.toLatestOppfolgingstilfelle
 import no.nav.syfo.dialogmote.api.domain.DialogmoteDTO
-import no.nav.syfo.dialogmote.database.getMoteStatusEndretNotPublished
+import no.nav.syfo.dialogmote.database.repository.MoteStatusEndretRepository
 import no.nav.syfo.dialogmote.domain.*
 import no.nav.syfo.testhelper.ExternalMockEnvironment
 import no.nav.syfo.testhelper.UserConstants.ARBEIDSTAKER_FNR
@@ -55,6 +55,7 @@ class AvlysDialogmoteApiV2Spek : Spek({
                 behandlerDialogmeldingProducer = behandlerDialogmeldingProducer,
             )
             val esyfovarselProducerMock = mockk<EsyfovarselProducer>(relaxed = true)
+            val moteStatusEndretRepository = MoteStatusEndretRepository(database)
 
             val altinnMock = mockk<ICorrespondenceAgencyExternalBasic>()
             val altinnResponse = ReceiptExternal()
@@ -180,7 +181,7 @@ class AvlysDialogmoteApiV2Spek : Spek({
 
                             verify(exactly = 0) { behandlerDialogmeldingProducer.sendDialogmelding(any()) }
 
-                            val moteStatusEndretList = database.getMoteStatusEndretNotPublished()
+                            val moteStatusEndretList = moteStatusEndretRepository.getMoteStatusEndretNotPublished()
                             moteStatusEndretList.size shouldBeEqualTo 2
 
                             moteStatusEndretList.forEach { moteStatusEndret ->
@@ -467,7 +468,7 @@ class AvlysDialogmoteApiV2Spek : Spek({
                                 LocalDateTime.now().isBefore(newDialogmoteDTO.tidSted.tid)
                             isTodayBeforeDialogmotetid shouldBeEqualTo false
 
-                            val moteStatusEndretList = database.getMoteStatusEndretNotPublished()
+                            val moteStatusEndretList = moteStatusEndretRepository.getMoteStatusEndretNotPublished()
                             moteStatusEndretList.size shouldBeEqualTo 2
 
                             moteStatusEndretList.forEach { moteStatusEndret ->
